Fixed some minor build breaking issues.

+++ src/entry.c
@@ -14,7 +14,7 @@
 #include <stdio.h>
 #include <locale.h>
 
-int main(const size_t num_of_args, const char *args[]) {
+int main(const int num_of_args, const char *args[]) {
   (void)num_of_args; (void)args;
   setlocale(LC_ALL, "en_US.UTF-8");
   fprintf(stdout, "Hello, Butane!\n");
